Clean code in preparation for new method
Remove superfluous code, including comments and debuggers. Prepare to
try two methods for populating current profiles after a successful
profile creation. The first will be a /GET request for the user after
the profile API updates to set app.user to current data. This feels like
the best approach. If this fails, I can attempt an /GET profiles#index
request and pass the returned data to the populator, but this seems like
an unclean solution.