Complete submittable 'change profiles' feature.
Get rounds associated with profile to clear when user changes profiles.
To accomplish this, I had to change the populateRounds function in the
profile.ui file. The change was adding a class 'profile-rounds' on the
table rows the function creates. Needed to do this so the empty format
row had a different identifier. Write function 'clearRounds' in
rounds.ui that sets html of elements with class 'profile-rounds' to
nothing.